Switched functionality to ESP32, refactored mobile app
