+++ MobileApp/TriadWatch/app/src/main/java/com/cnsprojectii/triadwatch/data/model/SensorData.kt
@@ -1,0 +1,8 @@
+package com.cnsprojectii.triadwatch.data.model
+
+@com.google.firebase.database.IgnoreExtraProperties
+data class SensorTypeData(
+    val latest: String? = null,
+    val hash: String? = null,
+    val timestamp: Long? = null // Assuming ESP32 sends UNIX timestamp in seconds
+)
