+++ MobileApp/TriadWatch/app/src/main/java/com/cnsprojectii/triadwatch/ui/state/HomeUIState.kt
@@ -1,0 +1,10 @@
+package com.cnsprojectii.triadwatch.ui.state
+
+data class TempHumidityUiState(
+    val temperature: String = "Loading...",
+    val humidity: String = "Loading...",
+    val isTemperatureVerified: Boolean = false,
+    val isHumidityVerified: Boolean = false,
+    val lastUpdateTimestamp: Long = 0L,
+    val statusMessage: String? = null
+)
